GEODE-5892: test rules should not modify the user.dir system property
* file's absolute path does not change anymore in jdk11 even after user.dir system property is changed.
