+++ geode-core/src/distributedTest/java/org/apache/geode/cache/query/dunit/PartitionedRegionCompactRangeIndexDUnitTest.java
@@ -40,7 +40,6 @@ import org.apache.geode.cache.query.data.Portfolio;
 import org.apache.geode.test.dunit.DistributedTestUtils;
 import org.apache.geode.test.dunit.IgnoredException;
 import org.apache.geode.test.dunit.SerializableRunnable;
-import org.apache.geode.test.dunit.VM;
 import org.apache.geode.test.dunit.rules.ClusterStartupRule;
 import org.apache.geode.test.dunit.rules.MemberVM;
 import org.apache.geode.test.junit.categories.OQLQueryTest;
@@ -58,9 +57,6 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
   MemberVM server1;
   MemberVM server2;
   MemberVM server3;
-  VM vm1;
-  VM vm2;
-  VM vm3;
 
   private Properties getSystemProperties(String cacheXML) {
     Properties props = new Properties();
@@ -77,9 +73,6 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     server1 = this.clusterStartupRule.startServerVM(1, props, this.locator.getPort());
     server2 = this.clusterStartupRule.startServerVM(2, props, this.locator.getPort());
     server3 = this.clusterStartupRule.startServerVM(3, props, this.locator.getPort());
-    vm1 = server1.getVM();
-    vm2 = server2.getVM();
-    vm3 = server3.getVM();
 
     // Adding due to known race condition for creation of partitioned indexes via cache.xml
     IgnoredException.addIgnoredException("IndexNameConflictException");
@@ -95,20 +88,20 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(i)));
 
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
-    vm2.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
+    server2.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
     clusterStartupRule.stop(2, false);
 
     // update entries
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
     clusterStartupRule.stop(1, false);
     server1 = this.clusterStartupRule.startServerVM(1, props, this.locator.getPort());
     server2 = this.clusterStartupRule.startServerVM(2, props, this.locator.getPort());
 
-    vm3.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
-    vm2.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
-    vm1.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
+    server3.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
+    server2.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
+    server1.invoke(verifyQueryResultsSize(idQuery, idQueryExpectedSize));
   }
 
   @Test
@@ -119,14 +112,14 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(10000 + i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
     // update entries
     IntStream.range(0, numEntries).forEach(i -> {
       entries.put("key-" + i, new Portfolio(i));
     });
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -142,15 +135,15 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
@@ -165,14 +158,14 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
     // update entries
     IntStream.range(0, numEntries).forEach(i -> {
       entries.put("key-" + i, new Portfolio(i));
     });
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -188,15 +181,15 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
@@ -211,7 +204,7 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(10000 + i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
 
@@ -219,7 +212,7 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     IntStream.range(0, numEntries).forEach(i -> {
       entries.put("key-" + i, new Portfolio(i));
     });
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -235,15 +228,15 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + ".entrySet where value.ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + ".entrySet where value.ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + ".entrySet where value.ID = ",
           () -> IntStream.range(0, numEntries), 8, 1);
     });
@@ -257,11 +250,11 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
 
-    vm3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
+    server3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -277,15 +270,15 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + ".entrySet where value.ID = ",
           () -> IntStream.range(0, numEntries), 8, 0);
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + ".entrySet where value.ID = ",
           () -> IntStream.range(0, numEntries), 8, 0);
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + ".entrySet where value.ID = ",
           () -> IntStream.range(0, numEntries), 8, 0);
     });
@@ -299,11 +292,11 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
 
-    vm3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
+    server3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -319,13 +312,13 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyIndexKeysAreEmpty();
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyIndexKeysAreEmpty();
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyIndexKeysAreEmpty();
     });
   }
@@ -337,10 +330,10 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
-    vm3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
+    server3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -356,15 +349,15 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 0);
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 0);
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyAllEntries("select key, value from /" + regionName + " where ID = ",
           () -> IntStream.range(0, numEntries), 8, 0);
     });
@@ -377,11 +370,11 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
     int numEntries = 100;
     Map<String, Portfolio> entries = new HashMap<>();
     IntStream.range(0, numEntries).forEach(i -> entries.put("key-" + i, new Portfolio(i)));
-    vm3.invoke(() -> populateRegion(regionName, entries));
+    server3.invoke(() -> populateRegion(regionName, entries));
 
     clusterStartupRule.stop(2, false);
 
-    vm3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
+    server3.invoke(() -> destroyFromRegion(regionName, entries.keySet()));
     clusterStartupRule.stop(1, false);
     clusterStartupRule.stop(3, false);
 
@@ -397,13 +390,13 @@ public class PartitionedRegionCompactRangeIndexDUnitTest implements Serializable
 
     // invoke the query enough times to hopefully randomize bucket to server targeting enough to
     // target both secondary/primary servers
-    vm3.invoke(() -> {
+    server3.invoke(() -> {
       verifyIndexKeysAreEmpty();
     });
-    vm2.invoke(() -> {
+    server2.invoke(() -> {
       verifyIndexKeysAreEmpty();
     });
-    vm1.invoke(() -> {
+    server1.invoke(() -> {
       verifyIndexKeysAreEmpty();
     });
   }
