+++ geode-core/src/integrationTest/java/org/apache/geode/management/internal/beans/ManagementAdapterTest.java
@@ -44,7 +44,7 @@ public class ManagementAdapterTest {
 
   @Rule
   public ServerStarterRule serverRule =
-      new ServerStarterRule().withWorkingDir().withLogFile().withAutoStart();
+      new ServerStarterRule().withLogFile().withAutoStart();
 
   @Rule
   public ConcurrencyRule concurrencyRule = new ConcurrencyRule();
