+++ geode-core/src/main/java/org/apache/geode/distributed/internal/ProductUseLog.java
@@ -53,7 +53,7 @@ public class ProductUseLog implements MembershipListener {
 
   public ProductUseLog(File productUseLogFile) {
     // GEODE-4180, use absolute paths
-    this.productUseLogFile = new File(productUseLogFile.getAbsolutePath());
+    this.productUseLogFile = productUseLogFile.getAbsoluteFile();
     this.logLevel = InternalLogWriter.INFO_LEVEL;
     createLogWriter();
   }
