+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/rules/ClusterStartupRule.java
@@ -381,6 +381,7 @@ public class ClusterStartupRule extends ExternalResource implements Serializable
 
   public static void stopElementInsideVM() {
     if (memberStarter != null) {
+      memberStarter.setCleanWorkingDir(false);
       memberStarter.after();
       memberStarter = null;
     }
