+++ geode-web/src/integrationTest/java/org/apache/geode/management/internal/cli/commands/CommandOverHttpTest.java
@@ -39,7 +39,7 @@ public class CommandOverHttpTest {
 
   @ClassRule
   public static ServerStarterRule server =
-      new ServerStarterRule().withWorkingDir().withLogFile().withJMXManager()
+      new ServerStarterRule().withLogFile().withJMXManager()
           .withHttpService()
           .withAutoStart();
